% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bastr_boxplot.R
\name{bastr_boxplot}
\alias{bastr_boxplot}
\title{Creates a boxplot in the style I like using ggplot}
\usage{
bastr_boxplot(
  x,
  var,
  group,
  title = NULL,
  stats = NULL,
  save.plot = F,
  path = getwd(),
  group.colors = NULL
)
}
\arguments{
\item{x}{(required) A dataframe containing the values to be plotted}

\item{var}{(required) The numeric variable to be plotted on y axis}

\item{group}{(required) The categorical variable to be plotted on x axis}

\item{title}{Title for plot}

\item{stats}{Data frame containing contrasts and their p-values}

\item{path}{Filepath for generated plots, default is working directory}

\item{saveplot}{True/False as to whether the plot should be saved to path}
}
\description{
Creates a boxplot in the style I like using ggplot
}
