% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bastr_lm.R
\docType{class}
\name{bastR_lm-class}
\alias{bastR_lm-class}
\title{This is a function for performing and summarizing linear models.
It is primarily meant to be used for generating many different models
with the same fixed/random effects and different response variables, e.g.,
for many different metabolites or cytokines. If there are no random
effects, this will use the lm() function. If there are random effects, this
will use the lmerTest() function.}
\arguments{
\item{x}{A data frame containing the predictors and response variables}

\item{vars}{A character vector specifying the names of the response variables to generate linear models for}

\item{fixed_effects}{A character vector specifying the names of the fixed effects. For example: c("group", "age")}

\item{random_effects}{A character vector specifying the random effects. For example: c("(1 | PID)")}

\item{check_model}{A boolean value specifying if linear model assumptions (e.g., linearity of residuals, heteroscedasticity, outliers) should be tested using the performance R package}
}
\description{
This is a function for performing and summarizing linear models.
It is primarily meant to be used for generating many different models
with the same fixed/random effects and different response variables, e.g.,
for many different metabolites or cytokines. If there are no random
effects, this will use the lm() function. If there are random effects, this
will use the lmerTest() function.
}
