% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bastr_lm.R
\name{bastr_lm}
\alias{bastr_lm}
\title{This is a function for performing and summarizing linear models.
It is primarily meant to be used for performing many different models
with the same fixed/random effects and different response variables, i.e.,
for many different metabolites or cytokines.}
\usage{
bastr_lm(x, vars, formula, check_model = T)
}
\description{
This is a function for performing and summarizing linear models.
It is primarily meant to be used for performing many different models
with the same fixed/random effects and different response variables, i.e.,
for many different metabolites or cytokines.
}
