% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Not in function
Returns the opposite of the \%in\% function
Returns true if an element is not in another vector, false if it is in it}
\usage{
... \%notin\% NA
}
\description{
Not in function
Returns the opposite of the \%in\% function
Returns true if an element is not in another vector, false if it is in it
}
