% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_zero_or_na}
\alias{is_zero_or_na}
\title{Return True/False depending on if value is zero or NA}
\usage{
is_zero_or_na(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A boolean vector of whether each index is a zero or NA
}
\description{
Return True/False depending on if value is zero or NA
}
