% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmheatmap.R
\name{lmheatmap}
\alias{lmheatmap}
\title{This function is for creating a mixed plot for visualizing many multiple
linear regression estimates for fixed effects. Creates line plots for
the primary fixed effect of interest and heatmaps for the estimates of
covariates}
\usage{
lmheatmap()
}
\arguments{
\item{x}{Data frame with the results from bastr_lm, or in the format of the output from broom_mixed of linear models}

\item{predictor}{Character string. Primary predictor to be used for showing line plot estimates}
}
\description{
This function is for creating a mixed plot for visualizing many multiple
linear regression estimates for fixed effects. Creates line plots for
the primary fixed effect of interest and heatmaps for the estimates of
covariates
}
