% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bastr_lm.R
\docType{class}
\name{MyObject-class}
\alias{MyObject-class}
\title{This is a function for performing and summarizing linear models.
It is primarily meant to be used for performing many different models
with the same fixed/random effects and different response variables, i.e.,
for many different metabolites or cytokines.}
\arguments{
\item{x}{A data frame containing the predictors and response variables}

\item{vars}{A character vector specifying the names of the response variables to generate linear models for}

\item{formula}{A character value specifying the formula of predictors to use for each response variable. Should include all fixed and random effects. Example: "age + bmi + (1|PID)"}

\item{check_model}{A boolean value specifying if linear model assumptions (e.g., linearity of residuals, heteroscedasticity, outliers) should be tested using the performance R package}
}
\description{
This is a function for performing and summarizing linear models.
It is primarily meant to be used for performing many different models
with the same fixed/random effects and different response variables, i.e.,
for many different metabolites or cytokines.
}
