% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{palettes}
\alias{palettes}
\title{Internal list of custom palettes for plots}
\format{
An object of class \code{list} of length 3.
}
\usage{
palettes
}
\description{
Internal list of custom palettes for plots
}
\keyword{internal}
