% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_missing.R
\name{replace_missing}
\alias{replace_missing}
\title{Replaces missing (NA) values in given columns of a dataframe.}
\usage{
replace_missing(x, cols, method = "lod2")
}
\arguments{
\item{x}{Dataframe containing the columns with missing values to be replaced}

\item{cols}{Vector of column names (strings) to have missing values replaced}

\item{method}{Method for replacing missing values. Default is lowest value in the column divided by 2}
}
\description{
Replaces missing (NA) values in given columns of a dataframe.
}
